move async io related code to linux os directory
This feature is only available in linux platform.
