+++ include/sys/dmu_impl.h
@@ -271,6 +271,8 @@ int dmu_write_direct(zio_t *, dmu_buf_impl_t *, abd_t *, dmu_tx_t *);
 int dmu_read_abd(dnode_t *, uint64_t, uint64_t, abd_t *, dmu_flags_t);
 int dmu_write_abd(dnode_t *, uint64_t, uint64_t, abd_t *, dmu_flags_t,
     dmu_tx_t *);
+abd_t *make_abd_for_dbuf(dmu_buf_impl_t *, abd_t *, uint64_t, uint64_t);
+void dmu_read_abd_done(zio_t *);
 #if defined(_KERNEL)
 int dmu_read_uio_direct(dnode_t *, zfs_uio_t *, uint64_t, dmu_flags_t);
 int dmu_write_uio_direct(dnode_t *, zfs_uio_t *, uint64_t, dmu_flags_t,
