+++ module/zfs/zfs_ioctl.c
@@ -8836,9 +8836,6 @@ zfs_kmod_init(void)
 
 	spa_init(SPA_MODE_READ | SPA_MODE_WRITE);
 	zfs_init();
-#ifdef __linux__
-	zfs_async_dio_init();
-#endif
 
 	zfs_ioctl_init();
 
@@ -8883,9 +8880,6 @@ zfs_kmod_fini(void)
 	}
 
 	zfs_ereport_taskq_fini();	/* run before zfs_fini() on Linux */
-#ifdef __linux__
-	zfs_async_dio_fini();
-#endif
 
 	/* Unregister zoned_uid callback before ZFS layer is torn down */
 	zone_unregister_zoned_uid_callback();
