+++ module/zfs/zfs_vnops.c
@@ -62,12 +62,6 @@
 #include <sys/zfs_vfsops.h>
 #include <sys/zfs_znode.h>
 
-#ifdef __linux__
-#include <sys/dmu_impl.h>
-#include <sys/zfs_vfsops_os.h>
-#include <sys/rrwlock.h>
-#endif
-
 /*
  * Enables access to the block cloning feature. If this setting is 0, then even
  * if feature@block_cloning is enabled, using functions and system calls that
@@ -102,7 +96,7 @@ static int zfs_dio_enabled = 1;
  * Strictly enforce alignment for Direct I/O requests, returning EINVAL
  * if not page-aligned instead of silently falling back to uncached I/O.
  */
-static int zfs_dio_strict = 0;
+int zfs_dio_strict = 0;
 
 
 /*
@@ -241,7 +235,7 @@ zfs_access(znode_t *zp, int mode, int flag, cred_t *cr)
  * following area for how this is handled:
  * zfs_write() -> update_pages()
  */
-static int
+int
 zfs_setup_direct(struct znode *zp, zfs_uio_t *uio, zfs_uio_rw_t rw,
     int *ioflagp)
 {
@@ -544,597 +538,8 @@ out:
 	return (error);
 }
 
-#ifdef __linux__
 
-/*
- * Dedicated caches for async read/write callback structs.
- * Eliminates per-I/O kmalloc fast-path overhead and improves
- * CPU cache locality by grouping same-type objects together.
- */
-static kmem_cache_t *zfs_async_read_cb_cache = NULL;
-static kmem_cache_t *zfs_async_write_cb_cache = NULL;
-
-/*
- * Completion callback for zfs_read_async().  Invoked from ZIO taskq
- * context when all blocks have been read.  Cleans up the range lock,
- * DIO pages, atime, and signals completion via kiocb->ki_complete().
- *
- * Linux-only: relies on struct kiocb and ki_complete().
- */
-struct zfs_async_read_cb
-{
-	struct kiocb	*kiocb;
-	znode_t		*zp;
-	zfsvfs_t	*zfsvfs;
-	zfs_locked_range_t *lr;
-	zfs_uio_t	uio;
-	ssize_t		start_resid;
-	boolean_t	dio;
-	const void	*tag;	/* saved FTAG for zfs_exit() */
-	int		lock_idx; /* saved sublock for teardown rrm */
-	abd_t		*data;	/* ABD wrapping user pages; freed in cb */
-};
-
-static void
-zfs_async_read_complete(void *arg, int error)
-{
-	struct zfs_async_read_cb *cb = arg;
-	ssize_t read;
-
-	/*
-	 * For the DIO (ABD) path, dmu_read_abd_async() writes data directly
-	 * to the user pages via DMA, so uio_resid is never decremented.
-	 * Report the full requested size as the bytes read.  On ECKSUM
-	 * retry via the ARC, uio_resid is decremented by zfs_uiomove, so
-	 * the subtraction works normally.
-	 */
-	if (cb->dio && error == 0)
-		read = cb->start_resid;
-	else
-		read = cb->start_resid - cb->uio.uio_resid;
-
-	/* ECKSUM retry via ARC */
-	if (error == ECKSUM && cb->dio) {
-		cb->uio.uio_extflg &= ~UIO_DIRECT;
-		cb->dio = B_FALSE;
-		dmu_buf_t *db = sa_get_db(cb->zp->z_sa_hdl);
-		error = dmu_read_uio_dnode(DB_DNODE((dmu_buf_impl_t *)db),
-		    &cb->uio, cb->start_resid, DMU_READ_PREFETCH);
-		if (error == 0)
-			read = cb->start_resid - cb->uio.uio_resid;
-		else
-			read = 0;
-	}
-
-	dataset_kstats_update_read_kstats(&cb->zfsvfs->z_kstat, read);
-	zfs_rangelock_exit(cb->lr);
-	if (cb->uio.uio_extflg & UIO_DIRECT)
-		zfs_uio_free_dio_pages(&cb->uio, UIO_READ);
-	ZFS_ACCESSTIME_STAMP(cb->zfsvfs, cb->zp);
-
-	/*
-	 * Release the teardown reader lock on the SAME sublock that was
-	 * acquired in zfs_read_async().  We cannot use zfs_exit() because
-	 * rrm_exit() hashes by curthread, and we are now running in ZIO
-	 * taskq context (different thread than the kworker that called
-	 * zfs_read_async).  Using the wrong sublock leaks the reader
-	 * count and causes zfsvfs_teardown to hang forever.
-	 */
-	zfs_exit_fs(cb->zfsvfs);
-	rrw_exit(&cb->zfsvfs->z_teardown_lock.locks[cb->lock_idx], cb->tag);
-
-	/*
-	 * Free the ABD that wraps the user pages.  The data is now in the
-	 * user pages, and the pages are unpinned by zfs_uio_free_dio_pages()
-	 * above.  In the synchronous path, the caller frees the ABD after
-	 * zio_wait() returns; here we must do it in the callback.
-	 */
-	if (cb->data != NULL)
-		abd_free(cb->data);
-
-#ifdef HAVE_2ARGS_KI_COMPLETE
-	cb->kiocb->ki_complete(cb->kiocb, error ? -error : read);
-#else
-	cb->kiocb->ki_complete(cb->kiocb, error ? -error : read, 0);
-#endif
-	kmem_cache_free(zfs_async_read_cb_cache, cb);
-}
-
-/*
- * Async Direct I/O read.  Pins user pages (requires current->mm),
- * submits reads via dmu_read_abd_async(), and returns 0 on success.
- * The caller should then return -EIOCBQUEUED to the VFS.
- *
- * On I/O completion, zfs_async_read_complete() handles cleanup and
- * calls kiocb->ki_complete().
- *
- * Returns 0 on successful async submission, or positive errno if the
- * read must be done synchronously.
- *
- * Linux-only: relies on struct kiocb, ki_complete(), -EIOCBQUEUED,
- * and the Linux VFS async I/O infrastructure.
- */
-int
-zfs_read_async(znode_t *zp, zfs_uio_t *uio, int ioflag, cred_t *cr,
-    struct kiocb *kiocb)
-{
-	(void) cr;
-	int error;
-	ssize_t n;
-
-	zfsvfs_t *zfsvfs = ZTOZSB(zp);
-	if ((error = zfs_enter_verify_zp(zfsvfs, zp, FTAG)) != 0)
-		return (error);
-
-	if (zp->z_pflags & ZFS_AV_QUARANTINED) {
-		zfs_exit(zfsvfs, FTAG);
-		return (SET_ERROR(EACCES));
-	}
-	if (Z_ISDIR(ZTOTYPE(zp))) {
-		zfs_exit(zfsvfs, FTAG);
-		return (SET_ERROR(EISDIR));
-	}
-	if (zfs_uio_offset(uio) < 0 || zfs_uio_resid(uio) == 0) {
-		zfs_exit(zfsvfs, FTAG);
-		return (zfs_uio_resid(uio) == 0 ? 0 : SET_ERROR(EINVAL));
-	}
-
-	/* Pin user pages for DIO — requires current->mm (we're in kworker) */
-	error = zfs_setup_direct(zp, uio, UIO_READ, &ioflag);
-	if (error || !(uio->uio_extflg & UIO_DIRECT)) {
-		zfs_exit(zfsvfs, FTAG);
-		return (error ? error : SET_ERROR(EOPNOTSUPP));
-	}
-
-	zfs_locked_range_t *lr = zfs_rangelock_enter(&zp->z_rangelock,
-	    zfs_uio_offset(uio), zfs_uio_resid(uio), RL_READER);
-
-	if (zfs_uio_offset(uio) >= zp->z_size) {
-		zfs_rangelock_exit(lr);
-		zfs_uio_free_dio_pages(uio, UIO_READ);
-		ZFS_ACCESSTIME_STAMP(zfsvfs, zp);
-		zfs_exit(zfsvfs, FTAG);
-		return (0);
-	}
-
-	n = MIN(zfs_uio_resid(uio), zp->z_size - zfs_uio_offset(uio));
-	ssize_t aligned_n = P2ALIGN_TYPED(n, PAGE_SIZE, ssize_t);
-	if (aligned_n == 0) {
-		zfs_rangelock_exit(lr);
-		zfs_uio_free_dio_pages(uio, UIO_READ);
-		zfs_exit(zfsvfs, FTAG);
-		return (SET_ERROR(EOPNOTSUPP));
-	}
-
-	/* Build ABD from pinned pages */
-	offset_t offset = zfs_uio_offset(uio);
-	offset_t page_idx = (offset - zfs_uio_soffset(uio)) >> PAGESHIFT;
-	ASSERT3U(page_idx, <, uio->uio_dio.npages);
-	abd_t *data = abd_alloc_from_pages(&uio->uio_dio.pages[page_idx],
-	    offset & (PAGESIZE - 1), aligned_n);
-
-	dmu_flags_t dflags = DMU_READ_PREFETCH | DMU_DIRECTIO;
-	if (ioflag & O_DIRECT)
-		dflags |= DMU_UNCACHEDIO;
-
-	struct zfs_async_read_cb *cb =
-	    kmem_cache_alloc(zfs_async_read_cb_cache, KM_SLEEP);
-	cb->kiocb = kiocb;
-	cb->zp = zp;
-	cb->zfsvfs = zfsvfs;
-	cb->lr = lr;
-	cb->uio = *uio;
-	cb->uio.uio_resid = aligned_n;
-	cb->start_resid = aligned_n;
-	cb->dio = B_TRUE;
-	cb->tag = FTAG;		/* saved for matching zfs_exit() in callback */
-	/*
-	 * Save the rrm sublock index.  zfs_enter_verify_zp() above took a
-	 * reader lock on z_teardown_lock using the kworker thread's pointer
-	 * for hashing.  The callback runs in ZIO taskq context (different
-	 * thread), so we must release the SAME sublock, not the one that
-	 * would be chosen by the callback's curthread.
-	 */
-	cb->lock_idx = ((uint32_t)(uintptr_t)(curthread)) % RRM_NUM_LOCKS;
-	cb->data = data;
-
-	dmu_buf_t *db = sa_get_db(zp->z_sa_hdl);
-	error = dmu_read_abd_async(DB_DNODE((dmu_buf_impl_t *)db),
-	    offset, aligned_n, data, dflags, zfs_async_read_complete, cb);
-
-	if (error) {
-		abd_free(data);
-		zfs_rangelock_exit(lr);
-		zfs_uio_free_dio_pages(uio, UIO_READ);
-		zfs_exit(zfsvfs, FTAG);
-		kmem_cache_free(zfs_async_read_cb_cache, cb);
-		return (error);
-	}
-
-	return (0);
-}
-static void zfs_clear_setid_bits_if_necessary(zfsvfs_t *, znode_t *, cred_t *,
-    uint64_t *, dmu_tx_t *);
-
-/*
- * Callback state for zfs_write_async().  Carries everything needed to
- * complete the write once the ZIOs finish.
- */
-struct zfs_async_write_cb {
-	struct kiocb	*kiocb;
-	znode_t		*zp;
-	zfsvfs_t	*zfsvfs;
-	zfs_locked_range_t *lr;
-	zfs_uio_t	uio;
-	ssize_t		start_resid;
-	boolean_t	dio;
-	const void	*tag;
-	int		lock_idx;
-	abd_t		*data;
-	dmu_tx_t	*tx;
-	dmu_flags_t	dflags;
-	offset_t	woff;
-	ssize_t		tx_bytes;
-	sa_bulk_attr_t	bulk[4];
-	int		bulk_count;
-	uint64_t	mtime[2];
-	uint64_t	ctime[2];
-	uint64_t	clear_setid_bits_txg;
-	cred_t		*cr;
-	boolean_t	do_commit;
-	int		error;
-	ssize_t		wrote;
-};
-
-static void
-zfs_async_write_task(void *arg)
-{
-	struct zfs_async_write_cb *cb = arg;
-	zilog_t *zilog = cb->zfsvfs->z_log;
-
-	if (cb->error == 0) {
-		if (cb->cr != NULL) {
-			zfs_clear_setid_bits_if_necessary(cb->zfsvfs, cb->zp,
-			    cb->cr, &cb->clear_setid_bits_txg, cb->tx);
-		}
-
-		zfs_tstamp_update_setup(cb->zp, CONTENT_MODIFIED,
-		    cb->mtime, cb->ctime);
-
-		uint64_t end_size;
-		while ((end_size = cb->zp->z_size) <
-		    (uint64_t)(cb->woff + cb->wrote))
-			(void) atomic_cas_64(&cb->zp->z_size, end_size,
-			    (uint64_t)(cb->woff + cb->wrote));
-
-		(void) sa_bulk_update(cb->zp->z_sa_hdl, cb->bulk,
-		    cb->bulk_count, cb->tx);
-
-		zfs_log_write(zilog, cb->tx, TX_WRITE, cb->zp, cb->woff,
-		    cb->wrote, cb->do_commit, cb->dio, NULL, NULL);
-
-		dmu_tx_commit(cb->tx);
-	} else {
-		dmu_tx_abort(cb->tx);
-		cb->wrote = 0;
-	}
-
-	dataset_kstats_update_write_kstats(&cb->zfsvfs->z_kstat, cb->wrote);
-	zfs_znode_update_vfs(cb->zp);
-	zfs_rangelock_exit(cb->lr);
-	if (cb->uio.uio_extflg & UIO_DIRECT)
-		zfs_uio_free_dio_pages(&cb->uio, UIO_WRITE);
-
-	/* zil_commit before teardown lock release (matches zfs_write()) */
-	if (cb->error == 0 && cb->do_commit) {
-		int zil_err = zil_commit(zilog, cb->zp->z_id);
-		if (zil_err != 0 && cb->error == 0)
-			cb->error = zil_err;
-	}
-
-	zfs_exit_fs(cb->zfsvfs);
-	rrw_exit(&cb->zfsvfs->z_teardown_lock.locks[cb->lock_idx], cb->tag);
-
-	if (cb->data != NULL)
-		abd_free(cb->data);
-	if (cb->cr != NULL)
-		crfree(cb->cr);
-
-#ifdef HAVE_2ARGS_KI_COMPLETE
-	cb->kiocb->ki_complete(cb->kiocb,
-	    cb->error ? -cb->error : cb->wrote);
-#else
-	cb->kiocb->ki_complete(cb->kiocb,
-	    cb->error ? -cb->error : cb->wrote, 0);
-#endif
-	kmem_cache_free(zfs_async_write_cb_cache, cb);
-}
-
-/*
- * Completion callback for dmu_write_abd_async().  Called from ZIO
- * taskq context when all writes finish.  Dispatches to system_taskq
- * for metadata operations, transaction commit, cleanup, and ki_complete
- * — all in process context, not ZIO taskq.
- */
-static void
-zfs_async_write_complete(void *arg, int error)
-{
-	struct zfs_async_write_cb *cb = arg;
-
-	if (cb->dio && error == 0)
-		cb->wrote = cb->tx_bytes;
-	else
-		cb->wrote = cb->start_resid - cb->uio.uio_resid;
-	cb->error = error;
-
-	taskqid_t tid = taskq_dispatch(system_taskq,
-	    zfs_async_write_task,
-	    cb, TQ_SLEEP);
-	if (tid == TASKQID_INVALID) {
-		/* Fallback: complete with error in current context */
-		cb->error = (error != 0) ? error : ENOMEM;
-		zfs_async_write_task(cb);
-	}
-}
-
-/*
- * Async Direct I/O write.  Pins pages, acquires locks, creates a
- * transaction, and dispatches writes via dmu_write_abd_async().
- * The kworker returns 0 immediately; the caller returns -EIOCBQUEUED.
- *
- * On I/O completion, zfs_async_write_complete() fires from ZIO taskq
- * and dispatches to system_taskq for metadata operations and cleanup
- * in process context.
- *
- * Linux-only: relies on struct kiocb, ki_complete(), -EIOCBQUEUED.
- */
-int
-zfs_write_async(znode_t *zp, zfs_uio_t *uio, int ioflag, cred_t *cr,
-    struct kiocb *kiocb)
-{
-	int error;
-	ssize_t n;
-
-	zfsvfs_t *zfsvfs = ZTOZSB(zp);
-	if ((error = zfs_enter_verify_zp(zfsvfs, zp, FTAG)) != 0)
-		return (error);
-
-	if (zp->z_pflags & ZFS_AV_QUARANTINED) {
-		zfs_exit(zfsvfs, FTAG);
-		return (SET_ERROR(EACCES));
-	}
-	if (Z_ISDIR(ZTOTYPE(zp))) {
-		zfs_exit(zfsvfs, FTAG);
-		return (SET_ERROR(EISDIR));
-	}
-	if (zfs_is_readonly(zfsvfs)) {
-		zfs_exit(zfsvfs, FTAG);
-		return (SET_ERROR(EROFS));
-	}
-	if ((zp->z_pflags & ZFS_IMMUTABLE) ||
-	    ((zp->z_pflags & ZFS_APPENDONLY) && !(ioflag & O_APPEND) &&
-	    (zfs_uio_offset(uio) < zp->z_size))) {
-		zfs_exit(zfsvfs, FTAG);
-		return (SET_ERROR(EPERM));
-	}
-
-	offset_t woff = ioflag & O_APPEND ?
-	    zp->z_size : zfs_uio_offset(uio);
-	if (woff < 0) {
-		zfs_exit(zfsvfs, FTAG);
-		return (SET_ERROR(EINVAL));
-	}
-
-	/*
-	 * Async Direct I/O writes copy user data into a kernel ABD
-	 * instead of pinning user pages via pin_user_pages_unlocked().
-	 * On RHEL kernels (and mainline >= 6.0), all pin_user_pages*()
-	 * variants implicitly add FOLL_LONGTERM which fails with ENOMEM
-	 * under concurrent I/O (iodepth=64 → 2048 concurrently held
-	 * pages).  Copying is safe on all kernel versions and the memcpy
-	 * cost is negligible compared to disk I/O latency.
-	 */
-	n = zfs_uio_resid(uio);
-	if (n == 0) {
-		zfs_exit(zfsvfs, FTAG);
-		return (0);
-	}
-
-	n = P2ALIGN_TYPED(n, PAGE_SIZE, ssize_t);
-	if (n == 0) {
-		zfs_exit(zfsvfs, FTAG);
-		return (SET_ERROR(EOPNOTSUPP));
-	}
-
-	/*
-	 * Replicate all DIO eligibility checks from zfs_setup_direct().
-	 * Any condition that would skip DIO in the sync path returns
-	 * EOPNOTSUPP so the caller falls back to zfs_write().
-	 */
-	if (zfsvfs->z_os->os_direct == ZFS_DIRECT_ALWAYS)
-		ioflag |= O_DIRECT;
-
-	if (!zfs_dio_enabled ||
-	    zfsvfs->z_os->os_direct == ZFS_DIRECT_DISABLED) {
-		zfs_exit(zfsvfs, FTAG);
-		return (SET_ERROR(EOPNOTSUPP));
-	}
-
-	if (!zfs_uio_page_aligned(uio) ||
-	    !zfs_uio_aligned(uio, PAGE_SIZE)) {
-		if (zfs_dio_strict)
-			return (SET_ERROR(EINVAL));
-		zfs_exit(zfsvfs, FTAG);
-		return (SET_ERROR(EOPNOTSUPP));
-	}
-
-	if (n < zp->z_blksz ||
-	    zn_has_cached_data(zp, zfs_uio_offset(uio),
-	    zfs_uio_offset(uio) + n - 1)) {
-		zfs_exit(zfsvfs, FTAG);
-		return (SET_ERROR(EOPNOTSUPP));
-	}
-
-	zfs_locked_range_t *lr;
-	if (ioflag & O_APPEND) {
-		lr = zfs_rangelock_enter(&zp->z_rangelock, 0, n, RL_APPEND);
-		woff = lr->lr_offset;
-		if (lr->lr_length == UINT64_MAX)
-			woff = zp->z_size;
-		zfs_uio_setoffset(uio, woff);
-	} else {
-		lr = zfs_rangelock_enter(&zp->z_rangelock, woff, n, RL_WRITER);
-	}
-
-	if (zn_rlimit_fsize_uio(zp, uio)) {
-		zfs_rangelock_exit(lr);
-		zfs_exit(zfsvfs, FTAG);
-		return (SET_ERROR(EFBIG));
-	}
-
-	if (lr->lr_length == UINT64_MAX) {
-		zfs_rangelock_exit(lr);
-		zfs_exit(zfsvfs, FTAG);
-		return (SET_ERROR(EOPNOTSUPP));
-	}
-
-	if (woff >= MAXOFFSET_T) {
-		zfs_rangelock_exit(lr);
-		zfs_exit(zfsvfs, FTAG);
-		return (SET_ERROR(EFBIG));
-	}
-	if ((uint64_t)n > MAXOFFSET_T - woff)
-		n = MAXOFFSET_T - woff;
-	n = MIN(n, zfs_uio_resid(uio));
-	n = P2ALIGN_TYPED(n, PAGE_SIZE, ssize_t);
-	if (n == 0) {
-		zfs_rangelock_exit(lr);
-		zfs_exit(zfsvfs, FTAG);
-		return (SET_ERROR(EOPNOTSUPP));
-	}
-
-	/*
-	 * Copy user data into a kernel ABD.  This avoids the FOLL_LONGTERM
-	 * pinning issues that plague the pin_user_pages*() family across
-	 * RHEL/mainline kernel versions.
-	 */
-	offset_t offset = zfs_uio_offset(uio);
-	abd_t *data = abd_alloc_linear(n, B_FALSE);
-	error = zfs_uiomove(abd_to_buf(data), n, UIO_WRITE, uio);
-	if (error) {
-		abd_free(data);
-		zfs_rangelock_exit(lr);
-		zfs_exit(zfsvfs, FTAG);
-		return (error);
-	}
-
-	boolean_t do_commit = (ioflag & (O_SYNC | O_DSYNC)) ||
-	    (zfsvfs->z_os->os_sync == ZFS_SYNC_ALWAYS);
-	dmu_tx_t *tx = dmu_tx_create(zfsvfs->z_os);
-	dmu_tx_hold_sa(tx, zp->z_sa_hdl, B_FALSE);
-	dmu_buf_impl_t *db = (dmu_buf_impl_t *)sa_get_db(zp->z_sa_hdl);
-	DB_DNODE_ENTER(db);
-	dmu_tx_hold_write_by_dnode(tx, DB_DNODE(db), woff, n);
-	DB_DNODE_EXIT(db);
-	zfs_sa_upgrade_txholds(tx, zp);
-	error = dmu_tx_assign(tx, DMU_TX_WAIT);
-	if (error) {
-		dmu_tx_abort(tx);
-		abd_free(data);
-		zfs_rangelock_exit(lr);
-		zfs_exit(zfsvfs, FTAG);
-		return (error);
-	}
-
-	dmu_flags_t dflags = DMU_READ_PREFETCH | DMU_DIRECTIO;
-	if (ioflag & O_DIRECT)
-		dflags |= DMU_UNCACHEDIO;
-
-	struct zfs_async_write_cb *cb =
-	    kmem_cache_alloc(zfs_async_write_cb_cache, KM_SLEEP);
-	cb->kiocb = kiocb;
-	cb->zp = zp;
-	cb->zfsvfs = zfsvfs;
-	cb->lr = lr;
-	cb->uio = *uio;
-	cb->uio.uio_resid = 0;		/* all data already copied to ABD */
-	cb->start_resid = n;
-	cb->dio = B_FALSE;
-	cb->tag = FTAG;
-	cb->lock_idx = ((uint32_t)(uintptr_t)(curthread)) % RRM_NUM_LOCKS;
-	cb->data = data;
-	cb->tx = tx;
-	cb->dflags = dflags;
-	cb->woff = woff;
-	cb->tx_bytes = n;
-	cb->cr = cr;
-	crhold(cr);
-
-	cb->bulk_count = 0;
-	SA_ADD_BULK_ATTR(cb->bulk, cb->bulk_count,
-	    SA_ZPL_MTIME(zfsvfs), NULL, &cb->mtime, 16);
-	SA_ADD_BULK_ATTR(cb->bulk, cb->bulk_count,
-	    SA_ZPL_CTIME(zfsvfs), NULL, &cb->ctime, 16);
-	SA_ADD_BULK_ATTR(cb->bulk, cb->bulk_count,
-	    SA_ZPL_SIZE(zfsvfs), NULL, &zp->z_size, 8);
-	SA_ADD_BULK_ATTR(cb->bulk, cb->bulk_count,
-	    SA_ZPL_FLAGS(zfsvfs), NULL, &zp->z_pflags, 8);
-
-	cb->clear_setid_bits_txg = 0;
-	cb->do_commit = do_commit;
-	cb->error = 0;
-	cb->wrote = 0;
-
-	error = dmu_write_abd_async(DB_DNODE(db), offset, n, data, dflags,
-	    tx, zfs_async_write_complete, cb);
-	if (error) {
-		crfree(cr);
-		abd_free(data);
-		zfs_rangelock_exit(lr);
-		dmu_tx_abort(tx);
-		zfs_exit(zfsvfs, FTAG);
-		kmem_cache_free(zfs_async_write_cb_cache, cb);
-		return (error);
-	}
-
-	return (0);
-}
-
-/*
- * Module init: create dedicated caches and taskq for async I/O.
- * Called once from zfs_kmod_init() — no lazy-init races.
- */
 void
-zfs_async_dio_init(void)
-{
-	zfs_async_read_cb_cache = kmem_cache_create("zfs_async_read_cb",
-	    sizeof (struct zfs_async_read_cb), 0,
-	    NULL, NULL, NULL, NULL, NULL, 0);
-	zfs_async_write_cb_cache = kmem_cache_create("zfs_async_write_cb",
-	    sizeof (struct zfs_async_write_cb), 0,
-	    NULL, NULL, NULL, NULL, NULL, 0);
-}
-
-/*
- * Module fini: destroy caches and taskq created by zfs_async_dio_init().
- * Called once from zfs_kmod_fini().
- */
-void
-zfs_async_dio_fini(void)
-{
-	if (zfs_async_read_cb_cache != NULL) {
-		kmem_cache_destroy(zfs_async_read_cb_cache);
-		zfs_async_read_cb_cache = NULL;
-	}
-	if (zfs_async_write_cb_cache != NULL) {
-		kmem_cache_destroy(zfs_async_write_cb_cache);
-		zfs_async_write_cb_cache = NULL;
-	}
-}
-#endif /* __linux__ */
-
-
-static void
 zfs_clear_setid_bits_if_necessary(zfsvfs_t *zfsvfs, znode_t *zp, cred_t *cr,
     uint64_t *clear_setid_bits_txgp, dmu_tx_t *tx)
 {
